timegraph.R: use segments instead of full h. rules
Segments now only take up the width of a week rather than span the
entire graph.

Grouping by week in the CSV is no long required, and will be ignored.
Weeks are now generated based on calendar time. This can lead to less
accurate ends on both sides of the graph if users start near the end of
a calendar week or end near the start of a week.